All Functions in Character Selection work now.
Create Delete Character 100%
Name Check 90%
SelectServer 100%
Select Characters 100%
Start Game Button 100%

+++ SunCommon/Entities/CharacterPosition.cs
@@ -11,9 +11,9 @@ namespace SunCommon.Entities
         public int Id { get; set; }
         public int Region { get; set; }
         public int Angle { get; set; }
-        public float LocationX { get; set; }
-        public float LocationY { get; set; }
-        public float LocationZ { get; set; }
+        public short LocationX { get; set; }
+        public short LocationY { get; set; }
+        public short LocationZ { get; set; }
     }
 }
